Support for reading the page map
The smaps file gives us the RSS for block memory allocations,
but this is a summary. If several j9segments are included in an
smap we don't know how much each contributes to RSS, so we have
to estimate, using a proportional approach: we consider their
RSS contribution to be proportional to their virtual size.
This commit improves the situations by adding teh ability to read
the pagemap file which shows which pages are in physical memory and
which are not. To use this feature we must be running as root
and the process whose footprint we analyze must still be running.

This commit also changes how parameters are passed to the program:
rather than assuming parameters are given in a precise order, we
added one letter option parameters:
"-j javacore_file -s smaps_file -c callsites_file -p PID"

Signed-off-by: Marius Pirvu <mpirvu@ca.ibm.com>